refactor: 데이터포매팅 기능 분리
